+++ activemodel/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 7.0.0.rc1 (December 06, 2021) ##
+
 *   Remove support to Marshal load Rails 5.x `ActiveModel::AttributeSet` format.
 
     *Rafael Mendonça França*